Expose max_edge_res parameter
Configurable edge deduplication distance. Without deduplication
non-manifold mesh intersections can result in the skins of the morphing
connection profiles. Higher values are more likely to work, but
obviously result in lower resolution meshes that are more prone to
poking through the higher resolution walls. Thus, this is exposed to
enable the user to find the lowest value for their model that passes CGAL.